+++ example/lib/pages/custom_icons_example.dart
@@ -51,7 +51,7 @@ class _CustomIconsExamplePageState extends State<CustomIconsExamplePage> {
       DataSource(
         type: DataSourceType.network,
         source:
-            "https://www.radiantmediaplayer.com/media/big-buck-bunny-360p.mp4",
+            "https://movietrailers.apple.com/movies/paramount/the-spongebob-movie-sponge-on-the-run/the-spongebob-movie-sponge-on-the-run-big-game_h720p.mov",
       ),
       autoplay: true,
     );