+++ example/lib/pages/disabled_buttons_example_page.dart
@@ -55,7 +55,7 @@ class _DisabledButtonsExampleState extends State<DisabledButtonsExample> {
       DataSource(
         type: DataSourceType.network,
         source:
-            "https://www.radiantmediaplayer.com/media/big-buck-bunny-360p.mp4",
+            "https://movietrailers.apple.com/movies/paramount/the-spongebob-movie-sponge-on-the-run/the-spongebob-movie-sponge-on-the-run-big-game_h720p.mov",
       ),
       autoplay: true,
     );