+++ example/lib/pages/player_with_header_page.dart
@@ -37,7 +37,7 @@ class _PlayerWithHeaderPageState extends State<PlayerWithHeaderPage> {
       DataSource(
         type: DataSourceType.network,
         source:
-            "https://www.radiantmediaplayer.com/media/big-buck-bunny-360p.mp4",
+            "https://movietrailers.apple.com/movies/paramount/the-spongebob-movie-sponge-on-the-run/the-spongebob-movie-sponge-on-the-run-big-game_h720p.mov",
       ),
       autoplay: true,
     );